invoice_detector/requirements.txt
absl-py==2.1.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
distro==1.9.0
easyocr==1.7.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-restx==1.3.0
flatbuffers==24.3.25
fsspec==2024.2.0
gast==0.5.4
gevent==24.2.1
geventhttpclient==2.0.12
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
lazy_loader==0.3
libclang==18.1.1
locust==2.24.1
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.25.2
oauthlib==3.2.2
openai==1.12.0
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
roundrobin==0.0.4
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sklearn-pandas==2.2.0
sniffio==1.3.1
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
zope.event==5.0
zope.interface==6.2